consolidate EntityCards
