+++ app/dev/components/component-library.tsx
@@ -12,7 +12,6 @@ import {
 } from '@/components/ui/select'
 
 import { BadgeComponents } from './sets/badges'
-import { EntityCardSet } from './sets/entity-card'
 import { OREntityComboboxSet } from './sets/or-entity-combobox'
 
 type ComponentSet = {
@@ -24,10 +23,6 @@ export function ComponentLibrary() {
   const [selectedSet, setSelectedSet] = useQueryState('set', parseAsString.withDefault(''))
 
   const componentSets: ComponentSet[] = [
-    {
-      name: 'entity cards',
-      component: EntityCardSet,
-    },
     {
       name: 'badges',
       component: BadgeComponents,
